+++ NEWS.md
@@ -5,6 +5,7 @@
 # Version 1.0.2
 - Bug correction on the use of organic matter in Saxton (2006) equations (thanks to Milan Fischer).
 - Recodification of soilgridParams due to new SoilGrids REST API (removed dependency from GSIF).
+- New function 'redefineSoilLayers'
 
 # Version 1.0.1
 - New root functions